requirements.txt
numpy
PyQt5
PyOpenGL